+++ tests/test_cell_level_io.py
@@ -3,26 +3,39 @@ import numpy as np
 from ccramic.app.io.annotation_outputs import *
 import os
 import tempfile
-
+import pytest
+import platform
 
 def test_output_annotations_masks():
     with tempfile.TemporaryDirectory() as tmpdirname:
         annotations_dict = {'Patient1+++slide0+++pos1_1': {(('xaxis.range[0]', 384.3802395209581),
                                                         ('xaxis.range[1]', 487.6736526946108),
                                                         ('yaxis.range[0]', 426.1467065868263),
-                                                        ('yaxis.range[1]', 322.8532934131736)): {'title': 'test',
-                                                                                                 'body': 'test',
-                                                                                                 'cell_type': 'cell type 1',
-                                                                                                 'imported': False,
-                                                                                                 'annotation_column': 'ccramic_cell_annotation',
-                                                                                                 'type': 'zoom',
-                                                                                                 'channels': ['Ho165'],
-                                                                                                 'use_mask': None,
-                                                                                                 'mask_selection': None,
-                                                                                                 'mask_blending_level': 35,
-                                                                                                 'add_mask_boundary': [
-                                                                                                     ' add boundary']},
-                                                       'M216.41616766467067,157.58383233532933L235.27844311377245,185.42814371257487L240.6676646706587,210.57784431137725L241.56586826347305,239.32035928143713L241.56586826347305,254.58982035928145L233.48203592814372,270.75748502994014L207.43413173652695,293.2125748502994L189.47005988023952,299.5L161.625748502994,297.7035928143713L143.66167664670658,290.5179640718563L129.29041916167665,275.248502994012L119.41017964071857,256.3862275449102L117.61377245508982,224.94910179640718L132.88323353293413,188.12275449101796L143.66167664670658,186.32634730538922L174.2005988023952,185.42814371257487L179.58982035928145,166.56586826347305L184.0808383233533,154.88922155688624L185.87724550898204,153.99101796407186Z': {
+                                                        ('yaxis.range[1]', 322.8532934131736)):
+                                                        {'title': 'test',
+                                                        'body': 'test',
+                                                        'cell_type': 'cell type 1',
+                                                        'imported': False,
+                                                        'annotation_column': 'ccramic_cell_annotation',
+                                                        'type': 'zoom',
+                                                        'channels': ['Ho165'],
+                                                        'use_mask': None,
+                                                        'mask_selection': None,
+                                                        'mask_blending_level': 35,
+                                                        'add_mask_boundary': [' add boundary']},
+                                                       'M216.41616766467067,157.58383233532933L235.27844311377245,'
+                                                       '185.42814371257487L240.6676646706587,'
+                                                       '210.57784431137725L241.56586826347305,239.32035928143713L'
+                                                       '241.56586826347305,254.58982035928145L233.48203592814372,'
+                                                       '270.75748502994014L207.43413173652695,293.2125748502994L'
+                                                       '189.47005988023952,299.5L161.625748502994,297.7035928143713L'
+                                                       '143.66167664670658,290.5179640718563L129.29041916167665,'
+                                                       '275.248502994012L119.41017964071857,256.3862275449102L'
+                                                       '117.61377245508982,224.94910179640718L132.88323353293413,'
+                                                       '188.12275449101796L143.66167664670658,186.32634730538922L'
+                                                       '174.2005988023952,185.42814371257487L179.58982035928145,'
+                                                       '166.56586826347305L184.0808383233533,154.88922155688624L'
+                                                       '185.87724550898204,153.99101796407186Z': {
                                                            'title': 'test', 'body': 'test', 'cell_type': 'cell type 2',
                                                            'imported': False,
                                                            'annotation_column': 'ccramic_cell_annotation',
@@ -43,9 +56,12 @@ def test_output_annotations_masks():
 
         assert not os.path.exists(os.path.join(tmpdirname, "annotation_masks.zip"))
         output_dir = export_annotations_as_masks(annotations_dict, tmpdirname,
-                                                 'Patient1+++slide0+++pos1_1', (600, 600, 1))
+                                                 'Patient1+++slide0+++pos1_1', (600, 600))
         assert os.path.exists(os.path.join(output_dir))
 
+
+@pytest.mark.skipif(platform.system() == 'Windows',
+                    reason="Skip the test output of point annotations CSV in Windows (different base64 format)")
 def test_output_point_annotations_as_csv():
     with tempfile.TemporaryDirectory() as tmpdirname:
         annotations_dict = {'Patient1+++slide0+++pos1_1': {